Update brew defaults.
Added a function called briu to quickly install a formulae
and dump the bundle changes automatically. So, then I only
have to push the changes to main.

Fixed a issue in update_brewfile function. Now it will move
the callee to starting directory properly.

Fixed a issue in install_homebrew function. Now it can properly
skip if brew is already installed.